+++ index.conf.js
@@ -1,5 +1,5 @@
 exports.PROGRAM_NAME = "scp-sync-to-local";
-exports.PROGRAM_VERSION = "0.10.0";
+exports.PROGRAM_VERSION = "0.6.0";
 exports.PASSWORD_MASK = "**********";
 
 exports.DEFAULT_SKIPIFEXISTS = true;
@@ -9,3 +9,4 @@ exports.DEFAULT_REMOTEPATH = ".";
 exports.DEFAULT_LOCALPATH = ".";
 exports.DEFAULT_VERBOSE = false;
 exports.DEFAULT_QUIET = false;
+exports.DEFAULT_HOSTNAME = "localhost";